requirements.txt
# requirements.txt
st-supabase-connection==2.0.1
plotly_express==0.4.1
pandas==2.2.3
streamlit_autorefresh==1.0.1
